QueryBuilder is added
