+++ src/main/java/com/ecommerce/ecommApp/products/ElasticSearchUtil.java
@@ -45,7 +45,7 @@ public final class ElasticSearchUtil {
             String productJson = objectMapper.writeValueAsString(product);
             String jsonBody = String.format("{\"doc\":%s}", productJson);
             String endpoint = String.format("%s/%s/%s/%s", INET_ADDRESS, _INDEX, "_update", product.getProductId());
-            System.out.println(Communication.sendHttpRequest(endpoint, jsonBody, RequestMethod.POST));
+            Communication.sendHttpRequest(endpoint, jsonBody, RequestMethod.POST);
             log.info("Data updated in elastic seach for product {}", product.getProductId());
             return true;
         } catch (JsonProcessingException ex) {
@@ -107,15 +107,12 @@ public final class ElasticSearchUtil {
 //            product.setPrice(123);
 //            insertProduct(product);
 //        }
-//        System.out.println(insertProduct(product));
-////        product.setName("Ripped Jeans");
-////        System.out.println(updateProduct(product));
 //        deleteProduct(1);
 //        getAllProducts();
 //        searchProduct();
         String json = "{\"search_text\":\"a\",\"price\":{\"lte\":1,\"gte\":13},\"category\":\"abc\"}";
         QueryBuilder queryBuilder = new QueryBuilder(json);
-        System.out.println(queryBuilder.build());
+        queryBuilder.build();
     }
 
     private static List<Product> searchProduct() {
@@ -129,54 +126,5 @@ public final class ElasticSearchUtil {
         } catch (Exception e) {
             return null;
         }
-    }
-
-}
-
-class QueryBuilder {
-
-    private JSONObject jsonQuery = new JSONObject("{\"query\":{\"bool\":{}}}");
-    private JSONObject jsonObject;
-
-    public QueryBuilder(String json) throws JSONException {
-        jsonObject = new JSONObject(json);
-    }
-
-    public String build() throws JSONException {
-
-        if (jsonObject.has("price")) {
-            JSONObject range = jsonObject.getJSONObject("price");
-            int gte = range.getInt("gte");
-            int lte = range.getInt("lte");
-            String filter = String.format("{\"range\":{\"price\":{\"gte\":%s,\"lte\":%s}}}", gte, lte);
-            JSONObject filterObject = new JSONObject(filter);
-            jsonQuery.getJSONObject("query").getJSONObject("bool").put("filter", filterObject);
-        }
-
-        if (jsonObject.has("category")) {
-            String categoryValue = jsonObject.getString("category");
-            String mustValue = String.format("[{\"match\":{\"category\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}}]", categoryValue);
-            JSONArray mustArray = new JSONArray(mustValue);
-            jsonQuery.getJSONObject("query").getJSONObject("bool").put("must", mustArray);
-        }
-
-        if (jsonObject.has("search_text")) {
-            String text = jsonObject.getString("search_text");
-            if (!jsonObject.has("category")) {
-                //no category object
-                String shouldValue = String.format("[{\"match\":{\"product_name\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}},{\"match\":{\"product_description\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}},{\"match\":{\"category\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}}]"
-                        , text, text, text);
-                JSONArray shouldArray = new JSONArray(shouldValue);
-                jsonQuery.getJSONObject("query").getJSONObject("bool").put("should", shouldArray);
-            } else {
-                //it has category
-                String shouldValue = String.format("[{\"match\":{\"product_name\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}},{\"match\":{\"product_description\":{\"query\":\"%s\",\"fuzziness\":\"AUTO\"}}}]"
-                        , text, text);
-                JSONArray shouldArray = new JSONArray(shouldValue);
-                jsonQuery.getJSONObject("query").getJSONObject("bool").put("should", shouldArray);
-            }
-            jsonQuery.getJSONObject("query").getJSONObject("bool").put("minimum_should_match", 1);
-        }
-        return jsonQuery.toString();
     }
 }
